Introduced factory method `new_plugin`
