+++ xrlint/operation.py
@@ -4,11 +4,8 @@ from dataclasses import dataclass
 from inspect import isclass, getdoc
 from typing import Any, Type, Callable
 
-from xrlint.util.codec import (
-    JsonSerializable,
-    MappingConstructible,
-    JsonValue,
-)
+from xrlint.util.constructible import MappingConstructible
+from xrlint.util.serializable import JsonSerializable, JsonValue
 from xrlint.util.importutil import import_value
 from xrlint.util.naming import to_kebab_case
