+++ xrlint/plugins/core/rules/__init__.py
@@ -1,13 +1,10 @@
 from xrlint.constants import CORE_PLUGIN_NAME
-from xrlint.plugin import Plugin
-from xrlint.plugin import PluginMeta
+from xrlint.plugin import new_plugin
 from xrlint.version import version
 
 
-plugin = Plugin(
-    meta=PluginMeta(
-        name=CORE_PLUGIN_NAME,
-        version=version,
-        ref="xrlint.plugins.core:export_plugin",
-    )
+plugin = new_plugin(
+    name=CORE_PLUGIN_NAME,
+    version=version,
+    ref="xrlint.plugins.core:export_plugin",
 )
